[skeleton] Add /boot entry into /etc/fstab
Let's add a /boot entry in fstab file in order to mount in one command
'mount /boot'.

It whould be usefull to edit uEnv.txt on v2r w/o using a Big Brother
(PC)

The entry is marked as noauto in order to avoid the partition damage in
case of power loss.